switched version to 64bit integer and set the DB copy upload to use Unix time for version

+++ objectstore/store_error.go
@@ -22,7 +22,7 @@ func InitialiseStoreError(ctx context.Context, backupConfig shared.ConfigBackup,
 	return result
 }
 
-func (object *StoreError) Upload(newDbRecord shared.BackedUpFileProperties, version int, backupJobsState shared.BackupJobsStateInterface,
+func (object *StoreError) Upload(newDbRecord shared.BackedUpFileProperties, version int64, backupJobsState shared.BackupJobsStateInterface,
 	metadata bool) (remoteVersion string, cancelled bool, err error) {
 	return "", false, fmt.Errorf("unsupported backend of type: '%s'", object.storeType)
 }
@@ -32,11 +32,11 @@ func (object *StoreError) GetStoreDetails() (StoreName string, StoreType string)
 	return object.storeName, object.storeType
 }
 
-func (object *StoreError) MarkDeleted(existingDbRecord shared.BackedUpFileProperties, version int, metadata bool) (remoteVersion string, cancelled bool, err error) {
+func (object *StoreError) MarkDeleted(existingDbRecord shared.BackedUpFileProperties, version int64, metadata bool) (remoteVersion string, cancelled bool, err error) {
 	return "", false, fmt.Errorf("unsupported backend of type: '%s'", object.storeType)
 }
 
-func (object *StoreError) Delete(path string, objType string, version int, remoteVersion string, metadata bool) error {
+func (object *StoreError) Delete(path string, objType string, version int64, remoteVersion string, metadata bool) error {
 	return fmt.Errorf("unsupported backend of type: '%s'", object.storeType)
 }
